break out install package tests
